Initial commit with Telegram bot files obnova

+++ internal/models/accounting.go
@@ -2,9 +2,10 @@ package models
 
 import "time"
 
+// AccountingRecord represents an accounting entry
 type AccountingRecord struct {
-	ID          int64     `json:"id"`
-	OrderID     int64     `json:"order_id"`
+	ID          int       `json:"id"`
+	OrderID     int       `json:"order_id"`
 	UserID      int64     `json:"user_id"`
 	Type        string    `json:"type"`
 	Amount      float64   `json:"amount"`
