+++ internal/models/message.go
@@ -2,8 +2,9 @@ package models
 
 import "time"
 
+// Message represents a chat message
 type Message struct {
-	ID         int64     `json:"id"`
+	ID         int       `json:"id"`
 	UserID     int64     `json:"user_id"`
 	OperatorID int64     `json:"operator_id"`
 	Message    string    `json:"message"`
