+++ internal/models/review.go
@@ -2,9 +2,10 @@ package models
 
 import "time"
 
+// Review represents a review for an order
 type Review struct {
-	ID        int64     `json:"id"`
-	OrderID   int64     `json:"order_id"`
+	ID        int       `json:"id"`
+	OrderID   int       `json:"order_id"`
 	UserID    int64     `json:"user_id"`
 	Rating    int       `json:"rating"`
 	Comment   string    `json:"comment"`
